ui: Split balance into component
With the goal to make it more visible in multi-page app
